backend/requirements.txt
fastapi
uvicorn
google-genai
supabase
python-dotenv
pydantic
